Flatten export subpackage and consolidate redundant abstractions
