+++ src/dr_wandb/__init__.py
@@ -1,20 +1,21 @@
-from dr_wandb.export import (
-    ExportEngine,
-    ExportManifest,
+from dr_wandb.config import (
     ExportMode,
     ExportRequest,
-    ExportState,
-    ExportSummary,
-    FetchMode,
-    HistoryPolicy,
-    HistoryPolicyContext,
-    HistoryRow,
+    HistorySelection,
     HistoryWindow,
-    RecordStore,
-    RunSnapshot,
-    StaticHistoryPolicy,
-    WandbRun,
+    SyncMode,
 )
+from dr_wandb.engine import ExportEngine
+from dr_wandb.history import HistoryRow
+from dr_wandb.results import ExportManifest, ExportSummary
+from dr_wandb.state import ExportState
+from dr_wandb.store import (
+    ExportStore,
+    iter_history_rows,
+    load_manifest,
+    load_run_snapshots,
+)
+from dr_wandb.wandb_run import RunSnapshot, WandbRun
 
 __all__ = [
     "ExportEngine",
@@ -22,14 +23,15 @@ __all__ = [
     "ExportMode",
     "ExportRequest",
     "ExportState",
+    "ExportStore",
     "ExportSummary",
-    "FetchMode",
-    "HistoryPolicy",
-    "HistoryPolicyContext",
     "HistoryRow",
+    "HistorySelection",
     "HistoryWindow",
-    "RecordStore",
     "RunSnapshot",
-    "StaticHistoryPolicy",
+    "SyncMode",
     "WandbRun",
+    "iter_history_rows",
+    "load_manifest",
+    "load_run_snapshots",
 ]
