--- src/dr_wandb/export/export_modes.py
@@ -1,13 +1,0 @@
-from __future__ import annotations
-
-from enum import StrEnum
-
-
-class ExportMode(StrEnum):
-    METADATA = "metadata"
-    HISTORY = "history"
-
-
-class FetchMode(StrEnum):
-    INCREMENTAL = "incremental"
-    FULL_RECONCILE = "full_reconcile"
