+++ src/dr_wandb/selection.py
@@ -1,10 +1,11 @@
+"""Select W&B runs to export: incremental (new + refresh non-terminal) or full list."""
+
 from __future__ import annotations
 
 from typing import Any
 
-from dr_wandb.export.export_request import ExportRequest
-from dr_wandb.export.export_state import ExportState, RunTrackingState
-from dr_wandb.export.export_modes import FetchMode
+from dr_wandb.config import ExportRequest, SyncMode
+from dr_wandb.state import ExportState, RunTrackingState
 
 TERMINAL_RUN_STATES = {"finished", "failed", "crashed", "killed"}
 RUN_BATCH_SIZE = 100
@@ -17,11 +18,11 @@ def select_runs(
     state: ExportState,
 ) -> list[Any]:
     if (
-        request.fetch_mode == FetchMode.FULL_RECONCILE
+        request.sync_mode == SyncMode.FULL_RECONCILE
         or state.max_created_at is None
     ):
         return list(
-            iter_runs(
+            _iter_runs(
                 api=api,
                 entity=request.entity,
                 project=request.project,
@@ -31,7 +32,7 @@ def select_runs(
 
     selected_runs: list[Any] = []
     seen_run_ids: set[str] = set()
-    for run in iter_runs(
+    for run in _iter_runs(
         api=api,
         entity=request.entity,
         project=request.project,
@@ -47,11 +48,10 @@ def select_runs(
     refresh_ids = [
         run_id
         for run_id, tracking in state.runs.items()
-        if should_refresh_tracking_state(tracking)
-        and run_id not in seen_run_ids
+        if _should_refresh(tracking) and run_id not in seen_run_ids
     ]
-    for batch in chunked(refresh_ids, RUN_BATCH_SIZE):
-        for run in iter_runs(
+    for batch in _chunked(refresh_ids, RUN_BATCH_SIZE):
+        for run in _iter_runs(
             api=api,
             entity=request.entity,
             project=request.project,
@@ -66,7 +66,7 @@ def select_runs(
     return selected_runs
 
 
-def iter_runs(
+def _iter_runs(
     *,
     api: Any,
     entity: str,
@@ -83,16 +83,11 @@ def iter_runs(
     )
 
 
-def chunked(values: list[str], size: int) -> list[list[str]]:
+def _chunked(values: list[str], size: int) -> list[list[str]]:
     return [
         values[index : index + size] for index in range(0, len(values), size)
     ]
 
 
-def should_refresh_tracking_state(tracking: RunTrackingState) -> bool:
+def _should_refresh(tracking: RunTrackingState) -> bool:
     return tracking.run_state not in TERMINAL_RUN_STATES
-
-
-def run_state(run: Any) -> str | None:
-    value = getattr(run, "state", None)
-    return str(value) if value is not None else None
