--- tests/test_export_paths.py
@@ -1,71 +1,0 @@
-from __future__ import annotations
-
-from pathlib import Path
-import pytest
-
-from dr_wandb.export.export_paths import ExportPaths
-from dr_wandb.export.export_manifest import ExportManifest
-from dr_wandb.export.export_modes import ExportMode
-from dr_wandb.export.export_state import ExportState
-
-
-def test_export_paths_from_name_and_root(tmp_path: Path) -> None:
-    paths = ExportPaths.from_name_and_root("moe_runs", tmp_path)
-
-    assert paths.name == "moe_runs"
-    assert paths.data_root == tmp_path
-    assert paths.export_dir == tmp_path / "moe_runs"
-    assert paths.manifest_path == tmp_path / "moe_runs" / "manifest.json"
-    assert paths.state_path == tmp_path / "moe_runs" / "state.json"
-
-
-def test_export_paths_build_run_and_history_paths(tmp_path: Path) -> None:
-    paths = ExportPaths.from_name_and_root("moe_runs", tmp_path)
-
-    assert paths.runs_path == tmp_path / "moe_runs" / "runs.jsonl"
-    assert paths.history_path == tmp_path / "moe_runs" / "history.jsonl"
-
-
-def test_export_paths_load_and_save_state(tmp_path: Path) -> None:
-    paths = ExportPaths.from_name_and_root("moe_runs", tmp_path)
-    state = ExportState(name="moe_runs", entity="ml-moe", project="moe")
-
-    paths.save_state(state)
-
-    loaded = paths.load_state(entity="ml-moe", project="moe")
-    assert loaded == state
-
-
-def test_export_paths_load_state_rejects_mismatched_identity(
-    tmp_path: Path,
-) -> None:
-    paths = ExportPaths.from_name_and_root("moe_runs", tmp_path)
-    other = ExportState(name="other_runs", entity="ml-moe", project="other")
-
-    paths.save_state(other)
-
-    with pytest.raises(ValueError, match="State identity does not match"):
-        paths.load_state(entity="ml-moe", project="moe")
-
-
-def test_export_paths_load_and_save_manifest(tmp_path: Path) -> None:
-    paths = ExportPaths.from_name_and_root("moe_runs", tmp_path)
-    manifest = ExportManifest(
-        name="moe_runs",
-        entity="ml-moe",
-        project="moe",
-        mode=ExportMode.METADATA,
-        created_at="2024-01-01T00:00:00+00:00",
-        updated_at="2024-01-01T00:00:00+00:00",
-        runs_path=str(paths.runs_path),
-        history_path=None,
-        run_count=1,
-        history_count=0,
-        selected_history_keys=None,
-        history_window=None,
-    )
-
-    paths.save_manifest(manifest)
-
-    loaded = paths.load_manifest()
-    assert loaded == manifest
